+++ contracts/src/helpers/TypeConverter.sol
@@ -6,8 +6,8 @@ library TypeConverter {
         bytes memory s = new bytes(42);
         s[0] = '0';
         s[1] = 'x';
-        for (uint i = 0; i < 20; i++) {
-            bytes1 b = bytes1(uint8(uint(uint160(x)) / (2**(8 * (19 - i)))));
+        for (uint256 i = 0; i < 20; i++) {
+            bytes1 b = bytes1(uint8(uint256(uint160(x)) / (2**(8 * (19 - i)))));
             bytes1 hi = bytes1(uint8(b) / 16);
             bytes1 lo = bytes1(uint8(b) - 16 * uint8(hi));
             s[2 * i + 2] = char(hi);
