+++ contracts/src/interfaces/IWorldID.sol
@@ -11,11 +11,11 @@ interface IWorldID {
     /// @param proof The zero-knowledge proof
     /// @dev  Note that a double-signaling check is not included here, and should be carried by the caller.
     function verifyProof(
-        uint root,
-        uint groupId,
-        uint signalHash,
-        uint nullifierHash,
-        uint externalNullifierHash,
-        uint[8] calldata proof
+        uint256 root,
+        uint256 groupId,
+        uint256 signalHash,
+        uint256 nullifierHash,
+        uint256 externalNullifierHash,
+        uint256[8] calldata proof
     ) external view;
 }
